pyproject.toml
[project]
name = "digital-footprint-analyzer"
version = "0.1.0"
description = "Digital Footprint Analyzer using crewAI"
authors = [{ name = "Alemhar", email = "alemhar@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.165.1,<1.0.0",
    "gradio",
    "python-dotenv",
    "openai",
    "langchain-google-genai",
    "langchain-openai"
]

[project.scripts]
dfa = "dfa.main:run"
run_crew = "dfa.main:run"
train = "dfa.main:train"
replay = "dfa.main:replay"
test = "dfa.main:test"
dfa-ui = "dfa.ui_app:main"
run-ui = "run_ui:main"

[project.optional-dependencies]
# Install these only if you need crewAI tools that may pull in heavy deps like chroma-hnswlib on Windows.
tools = [
    "crewai[tools]>=0.165.1,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dfa"]

[tool.hatch.build.targets.sdist]
include = [
  "src/dfa",
  "README.md",
  "pyproject.toml"
]

[tool.crewai]
type = "crew"
